Have a working ARFF exporter for the Parkinson datasets that import into Weka 3.8.2.
Updated the parser again to fix the ARFF exporter and quick fix for a label issue.
Issue was '%' character in labels for one of the columns. Converted to 'Percent' to fix quickly.

Todo: Need to add this to the p2a code for a check later. There is a parameter we can take advantage of for this.